+++ inmap/cmd/run.go
@@ -1,3 +1,21 @@
+/*
+Copyright © 2013 the InMAP authors.
+This file is part of InMAP.
+
+InMAP is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+InMAP is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with InMAP.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
 package cmd
 
 import "github.com/spf13/cobra"
@@ -22,7 +40,7 @@ func init() {
 			"gradients in population density and concentration.")
 	steadyCmd.PersistentFlags().BoolVar(&createGrid, "creategrid", false,
 		"Create the variable-resolution grid as specified in the configuration file"+
-			"before starting the simulation instead of reading it from a file. "+
+			" before starting the simulation instead of reading it from a file. "+
 			"If --dynamic is set to true, then this flag will also be automatically set to true.")
 
 }
@@ -30,14 +48,15 @@ func init() {
 var runCmd = &cobra.Command{
 	Use:   "run",
 	Short: "Run the model.",
-	Long:  "Run InMAP. Subcommands specify the run mode.",
+	Long: "run runs an InMAP simulation. Use the subcommands specified below to " +
+		" choose a run mode. (Currently 'steady' is the only avaible run mode.)",
 }
 
 // steadyCmd is a command that runs a steady-state simulation.
 var steadyCmd = &cobra.Command{
 	Use:   "steady",
 	Short: "Run InMAP in steady-state mode.",
-	Long: "Run InMAP in steady-state mode to calculate annual average " +
+	Long: "steady runs InMAP in steady-state mode to calculate annual average " +
 		"concentrations with no temporal variability.",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return Run(dynamic, createGrid)