+++ save_test.go
@@ -1,10 +1,26 @@
+/*
+Copyright © 2013 the InMAP authors.
+This file is part of InMAP.
+
+InMAP is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+InMAP is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with InMAP.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
 package inmap
 
 import (
 	"bytes"
 	"testing"
-
-	"github.com/ctessum/geom/index/rtree"
 )
 
 func TestSaveLoad(t *testing.T) {
@@ -12,9 +28,7 @@ func TestSaveLoad(t *testing.T) {
 	buf := bytes.NewBuffer([]byte{})
 
 	cfg, ctmdata, pop, popIndices, mr := VarGridData()
-	emis := &Emissions{
-		data: rtree.NewTree(25, 50),
-	}
+	emis := NewEmissions()
 
 	d := &InMAP{
 		InitFuncs: []DomainManipulator{